Reorganize grid drawing files
