+++ foundry/pixi/graphics/gridHighlight.d.ts
